final_project/pyproject.toml
[project]
name = "kafka-openfda-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Deepika-cloud-code", email = "vaddadi.d@northeastern.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = ["confluent-kafka (>=2.14.0,<3.0.0)", "requests (>=2.33.1,<3.0.0)", "python-dotenv (>=1.2.2,<2.0.0)", "snowflake-connector-python (>=4.4.0,<5.0.0)", "ijson (>=3.5.0,<4.0.0)", "datasets (>=4.8.4,<5.0.0)"]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"